requirements.txt
fastapi==0.103.0
uvicorn==0.23.2
xgboost==2.0.0
numpy==1.26.0
pandas==2.0.3
scikit-learn==1.3.0
pytest==7.4.2
pytest-cov==4.1.0
google-cloud-storage==2.10.0
python-dotenv==1.0.0
python-multipart==0.0.9
seaborn==0.13.0